requirements.txt
chainlit
torch
transformers
unsloth
Pillow
